set up server directory organization/subfolders, set up ts.config, start server.js, controller set up
