+++ package.json
@@ -7,5 +7,8 @@
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "Erik Gao, James Coen, Lauren Felty, Mike Bui, Piero Espejo",
-  "license": "ISC"
+  "license": "ISC",
+  "dependencies": {
+    "ts-node": "^10.9.2"
+  }
 }
