+++ server/package.json
@@ -17,7 +17,9 @@
     "bcrypt": "^5.1.1",
     "dotenv": "^16.4.5",
     "express": "^4.19.2",
-    "react-cookie": "^7.1.4"
+    "react-cookie": "^7.1.4",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.5.3"
   },
   "devDependencies": {
     "nodemon": "^3.1.4"
